Rename endpoints to handler
