--- endpoints/health.go
@@ -1,7 +1,0 @@
-package endpoints
-
-import "github.com/gin-gonic/gin"
-
-func Health(c *gin.Context) {
-	c.String(200, "pong")
-}
